#48 Split authenticateService to several services
