+++ Core/Interfaces/Email/EmailService.cs
@@ -1,9 +1,10 @@
-﻿using Microsoft.Extensions.Configuration;
+﻿using Core.Services.Email;
+using Microsoft.Extensions.Configuration;
 using System.Net;
 using System.Net.Mail;
 using System.Threading.Tasks;
 
-namespace Core.Services.Email
+namespace Core.Interfaces.Email
 {
     public class EmailService : IEmailService
     {
