+++ WebAPI/Controllers/AuthorsController.cs
@@ -1,17 +1,15 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Core.DTOs;
 using Core.Interfaces;
 using Swashbuckle.AspNetCore.Annotations;
 using Storage.Models;
 using Core.Requests;
+using Microsoft.AspNetCore.Authorization;
 
 namespace WebAPI.Controllers
 {
-    //[Authorize(Roles = "Admin")]
+    [Authorize(Roles = "Admin")]
     [Route("api/[controller]")]
     [ApiController]
     public class AuthorsController : ControllerBase
